requirements.txt
# ---------------------------------------------------------------------------
# Extra wheel index for CUDA 12.1 (must stay on top)
--extra-index-url https://download.pytorch.org/whl/cu121
# ---------------------------------------------------------------------------

# ---------- Core -----------------------------------------------------------
fastapi>=0.115.0
uvicorn[standard]>=0.24.0

jinja2>=3.1.2
pydantic>=2.3.0
python-dotenv>=1.0.0

# ---------- DB / Data ------------------------------------------------------
SQLAlchemy>=2.0.17
psycopg2-binary>=2.9.6
pandas>=2.0.3
esa>=1.3.0
redis>=4.5.5
pyarrow>=15.0.0            # Parquet engine for Pandas
pypdf>=4.2.0

# ---------- LLM stack ------------------------------------------------------
transformers==4.53.1
peft==0.16.0
sentence-transformers==5.0.0
accelerate>=0.28.0
bitsandbytes>=0.42.0
tokenizers>=0.15.2
faiss-cpu>=1.7.4
sentencepiece>=0.1.99
protobuf>=6.30.0,<7.0

# ---------- gRPC (LLM micro-service) ---------------------------------------
grpcio>=1.64.0          # runtime library
grpcio-tools>=1.73.0    # for `protoc --python_out= …`

# ---------- Utilities ------------------------------------------------------
python-multipart>=0.0.6
requests>=2.31.0
tabulate>=0.8.10
python-dateutil>=2.9.0

# ---------- PyTorch CUDA 12.1 wheels ---------------------------------------
torch==2.3.0+cu121
torchvision==0.18.0+cu121
torchaudio==2.3.0+cu121

# ---------- Critical pin: stay on NumPy 1.x for now ------------------------
numpy<2.0
